chat-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
mysql-connector-python==8.2.0
transformers==4.36.0
tokenizers==0.15.0
python-dotenv==1.0.0
